+++ __tests__/views/form.test.jsx
@@ -1,8 +1,5 @@
 import React from 'react';
 import Form from '../../src/views/form';
-import FormErrors from '../../src/views/formErrors';
-import Button from 'muicss/lib/react/button';
-// import Sinon from 'sinon';
 
 describe('Form', () =>{
     beforeEach(() => {
@@ -68,66 +65,5 @@ describe('<input />', () => {
         wrapper.instance().validateField('firstname', '!@#!@#$123456asdffadsfsad');
         expect (wrapper.state().firstnameValid).toBeFalsy();
     });
-    // it('Submitting the form', () => {
-    //     let wrapper = global.shallow(<Form />);
-    //     wrapper.setState({ firstname: 'Lauren', lastname: 'Chia', workphone: '03456789', mobile: '0123456789' });
-    //     console.log(wrapper.find('Button'));
-    //    wrapper.find('Button').simulate('click');
-    // });
+
 });
-
-describe('<Form Error />', () => {
-    it('`<Form error>` element should have a className', () => {
-        let wrapper = global.shallow(<FormErrors formErrors={{}}/>)
-        expect(
-            wrapper.find('div').hasClass('formErrors')
-        ).toBe(true);
-    });
-
-    it('`<Form error>` returns nothing if no errors are passed', () => {
-        let wrapper = global.mount(<FormErrors formErrors={{}}/>);
-        const errors = wrapper.find('.formErrors').map(node => node.text());
-        expect (errors).toEqual(['']);
-    });
-
-    // it('`<Form error>` returns errors if there is an error', () => {
-    //     let wrapper = global.shallow(<FormErrors formErrors={formErrors={firstname:'First'}}/>);
-    //     const errors = wrapper.find('.formErrors').map(node => node.text());
-    //     console.log(errors);
-    //     expect(wrapper.find('div.formErrors p')).toEqual(['']);
-    // });
-});
-
-    // it('should display an error when no value is input', () => {
-    //     const handleFormSubmit = spy();
-    //     wrapper = mount(<InputForm handleFormSubmit={handleFormSubmit} />);
-    //     wrapper.find('form').simulate('submit');
-    //     expect(
-    //         wrapper.state().fieldErrors.name
-    //     ).toBe('Please enter your name.');
-    // });
-
-
-	// it('handle changed input', () => {
-	// 	const wrapper = global.shallow(<Form />);
-	// 	wrapper.instance().changeValues = jest.fn();
-	// 	wrapper.update();
-	// 	wrapper.find('[label="firstname"]').simulate('change');
-	// 	expect(wrapper.instance().changeValues).toHaveBeenCalled();
-	// 	expect(wrapper).toMatchSnapshot();
-	// });
-	// //https://stackoverflow.com/questions/41830165/how-to-mock-react-component-methods-with-jest-and-enzyme
-    //
-	//
-	// it('test submit form', () => {
-	//     const wrapper = global.shallow(<Form />);
-	//     const preventDefault = jest.fn();
-	//     wrapper.find('form').simulate('click', { preventDefault });
-	//     expect(JSON.stringify(wrapper)).toMatchSnapshot();
-	//     expect(preventDefault).toBeCalled();
-  	// });
-	
-
-// //https://stackoverflow.com/questions/30034774/what-exactly-should-i-test-when-calling-reacts-this-props-onclickthis-in-jest
-
-// //https://github.com/airbnb/enzyme/issues/634
